+++ src/evaluation_system/settings/database.py
@@ -12,7 +12,6 @@ SETTINGS["INSTALLED_APPS"] = (
     "django.contrib.flatpages",
     "django.contrib.contenttypes",
     "django.contrib.sites",
-    "django_dbconn_retry",
 )
 SETTINGS["DATABASES"] = {
     "default": {